Decrease max distance Add toggle to target teammates Seperate target dummie and teammate toggles

+++ CRZAimbot/Main.h
@@ -20,8 +20,8 @@ void ProcessPlayer(Entity* LPlayer, Entity* target, UINT64 entitylist, int index
 void UpdatePlayersInfo(Entity* LocalPlayer);
 void PredictPosition(Entity* LocalPlayer, Entity* target, Vector* BonePosition);
 void AutoBoneSwitch();
-void SmoothType_Asist(float fov, float TargetDistance, Vector* Delta, float smooth_multiplier);
-void SmoothType_TargetLock(float fov, float TargetDistance, Vector* Delta, float smooth_multiplier);
+void SmoothType_Asist(float fov, float TargetDistance, Vector* Delta, int smooth_multiplier);
+void SmoothType_TargetLock(float fov, float TargetDistance, Vector* Delta, int smooth_multiplier);
 int AimAngles(Entity* LocalPlayer, Entity* target, Vector* out);
 void RunApp();
 void Configure();
